requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.7.1
pydantic-settings>=2.2.1
supabase>=2.4.5
python-dotenv>=1.0.1
pypdf>=4.2.0
python-docx>=1.1.2
anthropic>=0.18.0
